web_scraping/requirements.txt
azure-functions
requests
bs4
beautifulsoup4
mysql-connector-python
python-dotenv
urllib3
selenium
gitdb
GitPython
